// src/app/modules/land-creation/models/upload_documents.model.ts
export class UploadDocuments {
    LandInfoId: number;
    LandLatitude: string;
    LandLongitude: string;
    ImageBase64: string;
    VideoBase64: string;
    TranId: number;
    FileName: string;
    ImagePath: string;
    VideoPath: string;
    videofile: string;
    file: string;
    Path: string;
    ImageID: number;
}